Added filtering features for retrieving list of entities Warning => it is not a well-tested version !

+++ src/main/java/org/jco/spring/domain/crudify/ws/SpringCrudifyWsPage.java
@@ -2,7 +2,7 @@ package org.jco.spring.domain.crudify.ws;
 
 import java.util.List;
 
-public record SpringCrudifyWsPage(Integer totalCount, List<Object> entities) {
+public record SpringCrudifyWsPage(long totalCount, List<Object> entities) {
 
 
 }
